requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
Django==6.0.3
django-filter==25.2
django-rest-passwordreset==1.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.15
idna==3.11
inflection==0.5.1
packaging==26.0
PyJWT==2.12.1
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.5
uritemplate==4.2.0
urllib3==2.6.3
